A git-hook for detecting potentially missing files.
I frequently forget to "git add" new files.  This hook helps catch
the majority of cases by simply checking if any untracked files
appear within the diff of the potential commit.  This works for me,
as usually the missing file is a new source file - which I've diligently
added to the build system, but inevitably forget to git add.